triagent/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
anthropic>=0.28.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyGithub>=2.3.0
bcrypt>=4.1.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
httpx>=0.27.0
sse-starlette>=2.1.0
python-multipart>=0.0.9
numpy>=1.26.0
